Refactoring of hotels tests.
